--- src/pages/dashboard/DashboardPage.tsx
@@ -1,14 +1,0 @@
-import DashboardComponent from "../../components/dashboard/DashboardComponent";
-import BasicLayout from "../../layouts/BasicLayout.tsx";
-
-function DashboardPage() {
-    return (
-        <BasicLayout>
-            <div className="container mx-auto">
-                <DashboardComponent />
-            </div>
-        </BasicLayout>
-    );
-}
-
-export default DashboardPage;
